add theme typings
