+++ tsconfig.json
@@ -17,5 +17,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": ["src", "typings/worker-loader.d.ts"]
+  "include": ["src", "typings/worker-loader.d.ts", "typings/theme.d.ts"]
 }
