+++ typings/theme.d.ts
@@ -1,0 +1,5 @@
+import { Theme } from '@material-ui/core'
+
+declare module '@material-ui/styles/defaultTheme' {
+  interface DefaultTheme extends Theme {}
+}
